chore: pull core model/persistence theme support into the core
Fixes #2902
