+++ packages/core/i18n/resources_de.json
@@ -1,5 +1,6 @@
 {
   "today": "Heute um {0}",
+  "theme.unknown": "Unbekanntes Motiv",
   "sidecarExpand": "Auf volle Breite erweitern",
   "sidecarMinimize": "Minimieren",
   "sidecarClose": "Schließen",
