+++ packages/core/i18n/resources_fr.json
@@ -1,5 +1,6 @@
 {
   "today": "Aujourd'hui à {0}",
+  "theme.unknown": "Thème inconnu",
   "sidecarExpand": "Développer jusqu'à la pleine largeur",
   "sidecarMinimize": "Réduire",
   "sidecarClose": "Fermer",
