+++ packages/core/i18n/resources_it.json
@@ -1,5 +1,6 @@
 {
   "today": "Oggi alle {0}",
+  "theme.unknown": "Tema sconosciuto",
   "sidecarExpand": "Espandi a larghezza intera",
   "sidecarMinimize": "Riduci al minimo",
   "sidecarClose": "Chiudi",
