+++ packages/core/i18n/resources_ja.json
@@ -1,5 +1,6 @@
 {
   "today": "今日 {0}",
+  "theme.unknown": "不明なテーマ",
   "sidecarExpand": "全幅に展開",
   "sidecarMinimize": "最小化",
   "sidecarClose": "閉じる",
