+++ packages/core/i18n/resources_ko.json
@@ -1,5 +1,6 @@
 {
   "today": "{0}의 오늘",
+  "theme.unknown": "알 수 없는 테마",
   "sidecarExpand": "전체 너비로 확장",
   "sidecarMinimize": "최소화",
   "sidecarClose": "닫기",
