+++ packages/core/i18n/resources_pt_BR.json
@@ -1,5 +1,6 @@
 {
   "today": "Hoje às {0}",
+  "theme.unknown": "Tema desconhecido",
   "sidecarExpand": "Expandir para largura total",
   "sidecarMinimize": "Minimizar",
   "sidecarClose": "Fechar",
