+++ packages/core/i18n/resources_zh_CN.json
@@ -1,5 +1,6 @@
 {
   "today": "今天 {0}",
+  "theme.unknown": "未知主题",
   "sidecarExpand": "展开到完整宽度",
   "sidecarMinimize": "最小化",
   "sidecarClose": "关闭",
