+++ packages/core/i18n/resources_zh_TW.json
@@ -1,5 +1,6 @@
 {
   "today": "今天的 {0}",
+  "theme.unknown": "不明佈景主題",
   "sidecarExpand": "展開至整個寬度",
   "sidecarMinimize": "最小化",
   "sidecarClose": "關閉",
