+++ plugins/plugin-core-support/i18n/resources_de.json
@@ -31,7 +31,6 @@
   "Dark": "Dunkel",
   "Default Light": "Standard hell",
   "Default Dark": "Standard dunkel",
-  "theme.unknown": "Unbekanntes Motiv",
   "pleaseConfirm": "Bitte bestätigen",
   "areYouSure": "Sind Sie sicher?",
   "aboutToExecute": "Sie sind im Begriff, den Befehl auszuführen",
