+++ plugins/plugin-core-support/i18n/resources_en_US.json
@@ -31,7 +31,6 @@
   "Dark": "Dark",
   "Default Light": "Default Light",
   "Default Dark": "Default Dark",
-  "theme.unknown": "Unknown theme",
   "pleaseConfirm": "Please confirm",
   "areYouSure": "Are you sure?",
   "aboutToExecute": "You are about to execute the command",
