+++ plugins/plugin-core-support/i18n/resources_es.json
@@ -31,7 +31,6 @@
   "Dark": "Oscuro",
   "Default Light": "Claro predeterminado",
   "Default Dark": "Oscuro predeterminado",
-  "theme.unknown": "Tema desconocido",
   "pleaseConfirm": "Por favor, confirme",
   "areYouSure": "¿Está seguro?",
   "aboutToExecute": "Está a punto de ejecutar el mandato",
