+++ plugins/plugin-core-support/i18n/resources_fr.json
@@ -31,7 +31,6 @@
   "Dark": "Foncé",
   "Default Light": "Clair par défaut",
   "Default Dark": "Foncé par défaut",
-  "theme.unknown": "Thème inconnu",
   "pleaseConfirm": "Merci de confirmer",
   "areYouSure": "Confirmez-vous ?",
   "aboutToExecute": "Vous êtes sur le point d'exécuter la commande",
