+++ plugins/plugin-core-support/i18n/resources_it.json
@@ -31,7 +31,6 @@
   "Dark": "Scuro",
   "Default Light": "Chiaro predefinito",
   "Default Dark": "Scuro predefinito",
-  "theme.unknown": "Tema sconosciuto",
   "pleaseConfirm": "Confermare",
   "areYouSure": "Continuare l'operazione?",
   "aboutToExecute": "Si sta per eseguire il comando.",
