+++ plugins/plugin-core-support/i18n/resources_ja.json
@@ -31,7 +31,6 @@
   "Dark": "ダーク",
   "Default Light": "デフォルト・ライト",
   "Default Dark": "デフォルト・ダーク",
-  "theme.unknown": "不明なテーマ",
   "pleaseConfirm": "確認してください",
   "areYouSure": "実行しますか?",
   "aboutToExecute": "実行しようとしているコマンド:  ",
