+++ plugins/plugin-core-support/i18n/resources_ko.json
@@ -31,7 +31,6 @@
   "Dark": "다크",
   "Default Light": "기본 라이트",
   "Default Dark": "기본 다크",
-  "theme.unknown": "알 수 없는 테마",
   "pleaseConfirm": "확인 요망",
   "areYouSure": "계속하시겠습니까?",
   "aboutToExecute": "명령을 실행하려고 함",
