+++ plugins/plugin-core-support/i18n/resources_pt_BR.json
@@ -31,7 +31,6 @@
   "Dark": "Escuro",
   "Default Light": "Claro padrão",
   "Default Dark": "Escuro padrão",
-  "theme.unknown": "Tema desconhecido",
   "pleaseConfirm": "Confirme",
   "areYouSure": "Tem certeza?",
   "aboutToExecute": "Você está prestes a executar o comando",
