+++ plugins/plugin-core-support/i18n/resources_zh_CN.json
@@ -31,7 +31,6 @@
   "Dark": "深色",
   "Default Light": "默认浅色",
   "Default Dark": "默认深色",
-  "theme.unknown": "未知主题",
   "pleaseConfirm": "请确认",
   "areYouSure": "确定吗？",
   "aboutToExecute": "您即将执行命令",
