+++ plugins/plugin-core-support/i18n/resources_zh_TW.json
@@ -31,7 +31,6 @@
   "Dark": "暗",
   "Default Light": "預設亮",
   "Default Dark": "預設暗",
-  "theme.unknown": "不明佈景主題",
   "pleaseConfirm": "請確認",
   "areYouSure": "確定要執行嗎？",
   "aboutToExecute": "您即將執行指令",
