+++ plugins/plugin-core-support/src/preload.ts
@@ -29,12 +29,7 @@ const registration: Commands.PreloadRegistration = async (commandTree: Commands.
 
   if (!Capabilities.isHeadless()) {
     asyncs.push(import('./lib/cmds/zoom').then(_ => _.default(commandTree)))
-    asyncs.push(
-      import('./lib/cmds/theme')
-        .then(_ => _.preload())
-        .then(() => import('./lib/new-tab'))
-        .then(_ => _.default(commandTree))
-    )
+    asyncs.push(import('./lib/new-tab').then(_ => _.default(commandTree)))
     asyncs.push(import('./lib/cmds/history/reverse-i-search').then(_ => _.default()))
     asyncs.push(import('./lib/cmds/about/about').then(_ => _.preload()))
     asyncs.push(import('./lib/tab-completion').then(_ => _.default()))
